Backend/requirements.txt
flask
flask-cors
python-dotenv
openai
pymongo
dnspython
PyJWT
